working on authenticating sessionid

+++ static/js/admin.js
@@ -1,36 +1,45 @@
 const populateUserList = () => {
-    fetch("/getAllUserData")
+    fetch(`/getAllUserData?sessionid=${sessionid}`)
     .then((res) => {
         return res.json();
     })
     .then((res) => {
-        userlist = document.querySelector(".userList")
-        let template = "";
-        for (let i in res) {
-            devicelist = ""
-            for (let j in res[i].devices) {
-                devicelist += `<div class="device">${res[i].devices[j]}<div class="w3-button action" onclick="unlink('${res[i].username}', '${res[i].devices[j]}')">Huỷ liên kết</div></div>
-                        `;
+        if (res.status === "error") {
+            if (res.error === "expired") {
+                alert("Phiên đã hết hạn, vui lòng đăng nhập lại!");
+            } else if (res.error == "invalid") {
+                alert("Phiên không hợp lệ, vui lòng đăng nhập lại!");
             }
-            template += `<div class="user">
-                <div class="username-box">
-                    <p>Tên người dùng: </p>
-                    <p class="username">${res[i].username}</p>
-                    <div class="w3-button action" onclick="delUser('${res[i].username}')">Xoá</div>
-                </div>
-                <div class="device-list">
-                    <div class="device-action">
-                        <p>Các thiết bị đang liên kết: </p>
-                        <div class="w3-button action" onclick="newLink('${res[i].username}')">Liên kết mới</div>
-                    </div>
-                    <div class="devices">${devicelist}</div>
-                </div>
-            </div>`
         }
-        userlist.innerHTML = `        <div class="userList">
-            <div class="w3-button action" style="margin: 1rem 0rem;" onclick="createNewUser()">Tạo người dùng mới</div>
-            ${template}
-            </div>`;
+        else {
+            userlist = document.querySelector(".userList")
+            let template = "";
+            for (let i in res) {
+                devicelist = ""
+                for (let j in res[i].devices) {
+                    devicelist += `<div class="device">${res[i].devices[j]}<div class="w3-button action" onclick="unlink('${res[i].username}', '${res[i].devices[j]}')">Huỷ liên kết</div></div>
+                            `;
+                }
+                template += `<div class="user">
+                    <div class="username-box">
+                        <p>Tên người dùng: </p>
+                        <p class="username">${res[i].username}</p>
+                        <div class="w3-button action" onclick="delUser('${res[i].username}')">Xoá</div>
+                    </div>
+                    <div class="device-list">
+                        <div class="device-action">
+                            <p>Các thiết bị đang liên kết: </p>
+                            <div class="w3-button action" onclick="newLink('${res[i].username}')">Liên kết mới</div>
+                        </div>
+                        <div class="devices">${devicelist}</div>
+                    </div>
+                </div>`
+            }
+            userlist.innerHTML = `        <div class="userList">
+                <div class="w3-button action" style="margin: 1rem 0rem;" onclick="createNewUser()">Tạo người dùng mới</div>
+                ${template}
+                </div>`;
+        }
     })
 }
